requirements.txt
cffi==1.15.1
click==8.1.3
cryptography==38.0.1
ecdsa==0.18.0
Flask==2.2.2
Flask-Cors==3.0.10
Flask-SQLAlchemy==3.0.2
greenlet==1.1.3.post0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pyasn1==0.4.8
pycparser==2.21
python-dotenv==0.21.0
python-jose==3.3.0
rsa==4.9
six==1.16.0
SQLAlchemy==1.4.42
Werkzeug==2.2.2
